settings updated
#12342312321132

+++ twittur/forms.py
@@ -79,15 +79,19 @@ class UserDataForm(ModelForm):
 		# this is the current picture, nothing will happen if checkbox not clicked
 		picture = self.cleaned_data.get('picture')
 		# checkbox (False if clicked) -> return default picture
-		if picture == False:
+		if not picture:
 			return 'picture/default.gif'
 		return picture
 
 	def save(self, commit=True):
 		instance = super(UserDataForm, self).save(commit=False)
 		oldPic = self.oldPicture
-		if oldPic != 'picture/default.gif':
+		print(oldPic)
+		print(instance.picture)
+		if oldPic != 'picture/default.gif' and oldPic != instance.picture and oldPic != '':
 			oldPic.delete()
+			print(commit)
 		if commit:
+			print('im commit')
 			instance.save()
 		return instance
